asset: move unit test TestAssetGroupKey to group_key_test.go
